fix: copy db migrator to docker
