+++ apps/provider-inventory/package.json
@@ -11,7 +11,7 @@
     "dev": "npm run migration:exec && npm run start",
     "format": "prettier --write ./*.{ts,js,json} **/*.{ts,js,json}",
     "lint": "eslint .",
-    "migration:exec": "node ../../script/db-migrate.ts",
+    "migration:exec": "node dist/db-migrate.js",
     "migration:gen": "drizzle-kit generate",
     "prod": "node --permission --allow-fs-read=./dist/ --allow-fs-read=./env/ --allow-fs-read=./node_modules/ --allow-fs-read=../../node_modules/ --allow-fs-read=../../packages/env-loader/ --enable-source-maps --require ./dist/instrumentation.js ./dist/server.js",
     "start": "tsup --watch",
